refactor: 유틸, 에셋/아이콘 폴더 페이지 구분화
