+++ src/components/menu/Comment.tsx
@@ -1,5 +1,22 @@
 import React from 'react';
+import styled from 'styled-components';
 
 export default function Comment() {
-  return <div>Comment</div>;
+  return (
+    <CommentContainer>
+      <ProfileBox></ProfileBox>
+      <div>
+        <CommentInfoBox></CommentInfoBox>
+        <InteractingBox></InteractingBox>
+      </div>
+    </CommentContainer>
+  );
 }
+
+const CommentContainer = styled.div``;
+
+const ProfileBox = styled.div``;
+
+const CommentInfoBox = styled.div``;
+
+const InteractingBox = styled.div``;
