+++ src/pages/menu/MarketsPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { ReactComponent as SearchSVG } from '../../assets/icons/search.svg';
-import { ReactComponent as HeartSVG } from '../../assets/icons/heart.svg';
+import { ReactComponent as HeartSVG } from '../../assets/icons/menu/heart.svg';
 import { useNavigate } from 'react-router-dom';
 
 export default function MarketsPage() {
