+++ src/pages/menu/PostingDetailPage.tsx
@@ -2,17 +2,17 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import { Posting } from 'types/posting';
-import { getPostingDetail } from 'utils/getPostingDetail';
-import { ReactComponent as EyeSVG } from '../../assets/icons/eye.svg';
-import { ReactComponent as MoreSVG } from '../../assets/icons/more.svg';
-import { ReactComponent as LikeSVG } from '../../assets/icons/like.svg';
-import { postPostingViewCount } from 'utils/postPostingViewCount';
-import { postPostingLike } from 'utils/postPostingLike';
+import { getPostingDetail } from 'utils/menu/getPostingDetail';
+import { ReactComponent as EyeSVG } from '../../assets/icons/menu/eye.svg';
+import { ReactComponent as MoreSVG } from '../../assets/icons/menu/more.svg';
+import { ReactComponent as LikeSVG } from '../../assets/icons/menu/like.svg';
+import { postPostingViewCount } from 'utils/menu/postPostingViewCount';
+import { postPostingLike } from 'utils/menu/postPostingLike';
 import '../../../node_modules/react-quill-new/dist/quill.snow.css';
 import ReactQuill from 'react-quill-new';
-import { patchPosting } from 'utils/patchPosting';
-import { postComment } from 'utils/postComment';
-import { deletePosting } from 'utils/deletePosting';
+import { patchPosting } from 'utils/menu/patchPosting';
+import { postComment } from 'utils/menu/postComment';
+import { deletePosting } from 'utils/menu/deletePosting';
 
 export default function PostingDetailPage() {
   const location = useLocation();
@@ -179,6 +179,7 @@ export default function PostingDetailPage() {
             <p>
               Comment <span>{posting.comment_count}</span>
             </p>
+            <>{console.log(posting.comments)}</>
             <div>
               <textarea
                 ref={textarea}
