+++ src/types/comment.d.ts
@@ -1,0 +1,10 @@
+export interface Comment {
+  id: number;
+  user: number;
+  posting: number;
+  content: string;
+  created_at: string;
+  parent: number | null;
+  replies: Comment[];
+  like_count: number;
+}
