+++ src/utils/common/getPostingList.ts
